Add a unit test
This unit test was the first real test of the public
API, so it naturally ran into some issues larger
than the unit it was supposed to test.

  - Changed Screen from mut with lifetime to
    a Rc<RefCell<Screen>> so we can do compile
    time unsafe borrows (runtime will still panic
    if it is actually violating any rules)
  - Moved unit tests to bottom
  - Stressed out about endianess for too long

Hopefully the change to the smart pointer was
a good call. I feel it is best to avoid such things
if possible, but I can't think of a way to let two
references of Screen float around at once otherwise.
Obviously the Chip8 will have to have a mutable reference
to update the "VRAM", and a reference will have to float
around for the graphics api to access as well at the same
time. This breaks the rule of either 1 mutable reference,
or any number of immutable references at the same time at
compile time. This change circumvents that in what I believe
is a safe manner.