Finish instruction implementation
Finished all the instructions to be implemented. Note that this was done
quite lazily. Definitely better "rust" ways to do this versus the quick
and dirty c++ that I had implemented this with originally.

Removed all the weird ownership stuff I added previously. No need for
shared ownership in this situation. Just complicates things.

Added a few unit tests (not nearly enough!) and a helper function
to create test roms more easily.